Feat [#7] homeViewModel 생성 및 구현
화면 전환 구현 예정
