+++ DO_SOPT_33_Assignment/DO_SOPT_33_Assignment/Presentation/Home/ViewController/HomeViewController.swift
@@ -12,9 +12,6 @@ import Then
 
 final class HomeViewController: UIViewController {
     
-    var resultArray: [WeatherResponseDTO] = []
-    var mainWeathersData: [WeatherResponseDTO] = []
-    
     let moreButton = UIButton()
     let weatherTitleLabel = UILabel()
     let searchBar = UISearchBar()
@@ -23,9 +20,13 @@ final class HomeViewController: UIViewController {
                                                            collectionViewLayout: homeFlowLayout)
     private let homeFlowLayout = UICollectionViewFlowLayout()
     
+    var homeViewModel = HomeViewModel()
+
     override func viewDidLoad() {
         super.viewDidLoad()
         setUI()
+        bindViewModel()
+        setDelegate()
     }
     
     override func viewWillAppear(_ animated: Bool) {
@@ -39,11 +40,19 @@ final class HomeViewController: UIViewController {
         setCollectionViewConfig()
     }
     
+    private func setDelegate() {
+        homeViewModel.homeSearchBarDelegate = self
+    }
+    
+    private func bindViewModel() {
+        self.homeCollectionView.dataSource = homeViewModel
+        self.searchBar.delegate = homeViewModel
+    }
+    
     private func setCollectionViewConfig() {
         self.homeCollectionView.register(HomeWeatherCollectionViewCell.self,
                                          forCellWithReuseIdentifier: HomeWeatherCollectionViewCell.identifier)
         self.homeCollectionView.delegate = self
-        self.homeCollectionView.dataSource = self
     }
     
     private func setStyle() {
@@ -71,8 +80,6 @@ final class HomeViewController: UIViewController {
             
             // 위아래로 생기는 선 삭제
             $0.barTintColor = .clear
-            
-            $0.delegate = self
         }
         
         homeCollectionView.do {
@@ -119,46 +126,8 @@ final class HomeViewController: UIViewController {
     }
 }
 
-extension HomeViewController: WeatherButtonDelegate {
-    
-    func weatherButtonTapped(sender: WeatherListButton) {
-        let detailPageViewController = DetailPageViewController()
-        
-        for index in 0..<resultArray.count {
-            let detailViewController = DetailViewController()
-            detailViewController.indexNumber = index
-            detailViewController.detailWeatherData = resultArray[index]
-            detailPageViewController.detailViewControllers.append(detailViewController)
-        }
-        
-        let firstViewController = detailPageViewController.detailViewControllers[sender.indexNumber]
-        detailPageViewController.pageViewController.setViewControllers([firstViewController], direction: .forward, animated: true, completion: nil)
-        
-        detailPageViewController.detailViewControllers[sender.indexNumber].indexNumber = sender.indexNumber
-        self.navigationController?.pushViewController(detailPageViewController, animated: true)
-    }
-}
-
-extension HomeViewController: UISearchBarDelegate {
-    
-    func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
-        guard let text = searchBar.text else { return }
-        
-        self.resultArray = []
-        
-        mainWeathersData.forEach {
-            // 대소문자 구별은 안됨..
-            if $0.name.contains(text) {
-                resultArray.append($0)
-            }
-        }
-        
-        print(resultArray)
-        
-        if text.isEmpty {
-            self.resultArray = mainWeathersData
-        }
-        
+extension HomeViewController: HomeViewModelDelegate {
+    func updateData() {
         self.homeCollectionView.reloadData()
     }
 }
@@ -167,52 +136,17 @@ extension HomeViewController {
     override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?){
         self.view.endEditing(true)
     }
-}
-
-extension HomeViewController: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return resultArray.count
-    }
     
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: HomeWeatherCollectionViewCell.identifier, for: indexPath) as? HomeWeatherCollectionViewCell else { return UICollectionViewCell() }
-        cell.weatherButton.weatherButtonDelegate = self
-        cell.bindData(data: resultArray[indexPath.row])
-        cell.weatherButton.indexNumber = indexPath.row
-        return cell
+    private func loadWeatherData()  {
+        Task {
+            let success = await homeViewModel.loadWeatherData()
+            if success {
+                self.homeCollectionView.reloadData()
+                print("💛💛💛💛💛💛💛")
+            }
+        }
     }
 }
 
 extension HomeViewController: UICollectionViewDelegate { }
 
-extension HomeViewController {
-    private func loadWeatherData() {
-        Task {
-            do {
-                let cities = ["seoul", "daegu", "busan", "daejeon", "mokpo"]
-                
-                self.mainWeathersData = []
-                var weatherDataArray: [WeatherResponseDTO] = []
-                
-                for cityName in cities {
-                    do {
-                        if let receivedData = try await WeatherService.shared.GetWeatherData(cityName: cityName) {
-                            weatherDataArray.append(receivedData)
-                        }
-                    } catch {
-                        print("Failed to get weather data for \(cityName): \(error)")
-                    }
-                }
-
-                DispatchQueue.main.async {
-                    self.mainWeathersData = weatherDataArray
-                    self.resultArray = self.mainWeathersData
-                    self.homeCollectionView.reloadData()
-                    print("💛💛💛💛💛💛💛")
-                }
-            } catch {
-                print(error)
-            }
-        }
-    }
-}
